Refactored glitched logic functions for existing rules to new module.
