+++ worlds/rabi_ribi/client/patch.py
@@ -4,17 +4,17 @@ This is done on the client side upon connect to allow for a smoother setup exper
 """
 import os
 import struct
-from worlds.rabi_ribi.client.client import RabiRibiContext
-from worlds.rabi_ribi.data import data
-from worlds.rabi_ribi.existing_randomizer.dataparser import RandomizerData
-from worlds.rabi_ribi.existing_randomizer.mapfileio import (
+from .client import RabiRibiContext
+from ..data import data
+from ..existing_randomizer.dataparser import RandomizerData
+from ..existing_randomizer.mapfileio import (
     ItemModifier,
     grab_original_maps,
     MAP_ITEMS_OFFSET,
     MAP_TILES0_OFFSET,
     MAP_SIZE
 )
-from worlds.rabi_ribi.existing_randomizer.randomizer import (
+from ..existing_randomizer.randomizer import (
     apply_item_specific_fixes,
     apply_map_transition_shuffle,
     apply_start_location_shuffle,
@@ -23,11 +23,11 @@ from worlds.rabi_ribi.existing_randomizer.randomizer import (
     parse_args,
     pre_modify_map_data
 )
-from worlds.rabi_ribi.existing_randomizer.utility import to_index
-from worlds.rabi_ribi.items import lookup_item_id_to_name
-from worlds.rabi_ribi.locations import lookup_location_id_to_name
-from worlds.rabi_ribi.options import AttackMode
-from worlds.rabi_ribi.world import RabiRibiWorld
+from ..existing_randomizer.utility import to_index
+from ..items import lookup_item_id_to_name
+from ..locations import lookup_location_id_to_name
+from ..options import AttackMode
+from ..world import RabiRibiWorld
 
 class Allocation():
     """
